move version handler, rename plex handler
